Remove unnecessary files
