+++ src/App.js
@@ -1,8 +1,9 @@
+import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import Navbar from './components/Navbar';
 import store from './redux/configureStore';
-import AllBooks from './components/BookList';
+import AllBooks from './components/Books';
 import Categories from './components/Categories';
 
 function App() {
