+++ src/components/BookAdd.js
@@ -19,6 +19,7 @@ const BookAdd = () => {
     title,
     category,
   };
+
   const handleSubmit = (e) => {
     e.preventDefault();
     if (title && category) {
@@ -31,7 +32,7 @@ const BookAdd = () => {
 
   return (
     <section className="form-section">
-      <h2>Add a new book here</h2>
+      <h2>Add New Book</h2>
 
       <form className="form" onSubmit={handleSubmit}>
         <input
@@ -52,4 +53,5 @@ const BookAdd = () => {
     </section>
   );
 };
+
 export default BookAdd;
