+++ src/components/Categories.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 const Categories = () => (
   <div className="categories">
-    <h2>Under Construction</h2>
+    <h2>Under Construction...</h2>
   </div>
 );
