+++ src/index.css
@@ -1,0 +1,425 @@
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@300&display=swap');
+
+:root {
+  --pale-gray: #f5f6fa;
+  --white: #fff;
+  --azure: #0290ff;
+  --off-white: #e8e8e8;
+  --dark: #121212;
+  --light-blue: #4386bf;
+}
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  background-color: var(--pale-gray);
+  font-family: 'Montserrat', sans-serif;
+}
+
+header {
+  background-color: var(--white);
+  width: 100%;
+  position: fixed;
+  top: 0;
+}
+
+nav {
+  height: 4rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+nav,
+.list,
+hr {
+  width: 90%;
+  margin: 0 auto;
+}
+
+.navbar {
+  display: flex;
+  gap: 20px;
+  align-items: center;
+}
+
+.navbar h1 {
+  color: var(--azure);
+  font-weight: 700;
+  font-size: 1rem;
+}
+
+.navbar ul {
+  margin: 0;
+  padding: 0;
+}
+
+.navbar li {
+  list-style: none;
+  display: inline-block;
+  margin-right: 5px;
+}
+
+.nav-link {
+  font-size: 0.65rem;
+  text-decoration: none;
+  text-transform: uppercase;
+  letter-spacing: 1.9px;
+  color: var(--dark);
+}
+
+.current {
+  font-weight: 600;
+}
+
+.avatar {
+  padding: 5px 8px;
+  border: solid 1px var(--off-white);
+  border-radius: 50%;
+}
+
+.user {
+  color: var(--azure);
+}
+
+.no-books {
+  text-align: center;
+  margin-top: 150px;
+  line-height: 1.5;
+}
+
+.book-list {
+  width: 100%;
+  display: grid;
+  gap: 20px;
+  margin-top: 100px;
+}
+
+.list {
+  list-style: none;
+  height: auto;
+  border: solid 1px var(--off-white);
+  background-color: var(--white);
+  border-radius: 4px;
+  padding: 1.5rem;
+}
+
+.all-books,
+.description,
+.progress-chapter,
+.chapters,
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+}
+
+.all-books,
+.progress-chapter {
+  gap: 25px;
+}
+
+.book-info {
+  line-height: 1.6;
+}
+
+.book-info p {
+  font-size: 12px;
+}
+
+.category,
+.completed {
+  opacity: 0.5;
+  font-size: 0.875rem;
+  font-weight: bold;
+  color: var(--dark);
+}
+
+.completed {
+  font-weight: normal;
+}
+
+.title {
+  font-family: 'Roboto Slab', serif;
+  font-size: 1.375rem;
+  font-weight: bold;
+  color: var(--dark);
+  letter-spacing: -0.2px;
+}
+
+.author,
+.action button,
+.current-chapter-one,
+.current-chapter-two,
+.chapters button {
+  font-family: 'Roboto Slab', serif;
+  font-weight: 300;
+}
+
+.author,
+.action button {
+  color: var(--light-blue);
+}
+
+.current-chapter-one,
+.current-chapter-two {
+  font-size: 0.813rem;
+  color: var(--dark);
+  opacity: 0.5;
+}
+
+.current-chapter-two {
+  font-size: 1rem;
+  letter-spacing: -0.4px;
+  opacity: 1;
+}
+
+.action button {
+  display: inline-block;
+  margin-right: 13px;
+  border: none;
+  background-color: var(--white);
+  cursor: pointer;
+}
+
+.chapters button {
+  width: 10rem;
+  margin: 0 auto;
+  height: 2.063rem;
+  border-radius: 3px;
+  background-color: var(--azure);
+  color: var(--white);
+  border: none;
+  letter-spacing: 0.5px;
+  cursor: pointer;
+  transition: 300ms;
+}
+
+.action button:first-child,
+.action button:nth-child(2) {
+  padding-right: 8px;
+  border-right: 1px solid var(--off-white);
+}
+
+.action button:hover {
+  border-bottom: 1px solid var(--off-white);
+  color: var(--dark);
+}
+
+.progress,
+.current-chapter {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.progress-img {
+  width: 4rem;
+  height: 4rem;
+  border-radius: 50%;
+  border-color: #4386bf #4386bf #4386bf #e8e8e8;
+  border-style: solid;
+  border-width: 5px;
+  transform: rotate(45deg);
+}
+
+.vertical-line {
+  display: none;
+}
+
+.percent {
+  font-size: 2rem;
+  color: var(--dark);
+}
+
+.chapters button:hover {
+  opacity: 0.7;
+}
+
+hr {
+  margin-top: 40px;
+  border-top: 1px solid var(--off-white);
+}
+
+.form-section {
+  margin: 50px 0;
+  text-align: center;
+}
+
+.form-section h2 {
+  font-size: 1.25rem;
+  font-weight: bold;
+  letter-spacing: -0.18px;
+  color: #888;
+  padding-bottom: 25px;
+}
+
+.form {
+  width: 70%;
+  margin: 0 auto;
+}
+
+input,
+select,
+.submit-btn {
+  padding: 10px 20px;
+  border-radius: 4px;
+  border: solid 1px var(--white);
+  outline: none;
+}
+
+input::placeholder,
+select {
+  color: #c4c4c4;
+  font-size: 1rem;
+}
+
+.submit-btn {
+  background-color: var(--azure);
+  font-family: 'Roboto Slab', serif;
+  color: var(--white);
+  letter-spacing: 0.5px;
+  text-align: center;
+  text-transform: uppercase;
+}
+
+.categories {
+  height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin: 0 auto;
+}
+
+.categories h2 {
+  font-style: italic;
+  color: red;
+  font-family: cursive;
+}
+
+@media all and (min-width: 1024px) {
+  nav,
+  .list,
+  hr,
+  .form-section {
+    width: 85%;
+    margin: 0 auto;
+  }
+
+  nav {
+    height: 6rem;
+  }
+
+  .navbar {
+    gap: 50px;
+  }
+
+  .navbar h1 {
+    font-size: 1.875rem;
+  }
+
+  .navbar li {
+    margin-right: 35px;
+  }
+
+  .nav-link {
+    font-size: 0.813rem;
+  }
+
+  .no-books {
+    margin-top: 250px;
+    font-size: 1.7rem;
+  }
+
+  .book-list {
+    margin-top: 170px;
+  }
+
+  .all-books,
+  .progress-chapter,
+  .form {
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+  }
+
+  .book-info h3 {
+    font-size: 22px;
+    width: 20rem;
+  }
+
+  .book-info p {
+    font-size: 14px;
+  }
+
+  .progress-chapter {
+    gap: 40px;
+  }
+
+  .vertical-line {
+    display: block;
+    border: 1px solid var(--off-white);
+    width: 60px;
+    transform: rotate(90deg);
+  }
+
+  .progress {
+    gap: 20px;
+  }
+
+  .current-chapter {
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: center;
+    gap: 8px;
+  }
+
+  .chapters {
+    justify-content: space-between;
+  }
+
+  hr {
+    margin-top: 70px;
+  }
+
+  .form-section {
+    margin: 85px auto;
+  }
+
+  .form-section h2 {
+    font-size: 1.4rem;
+    text-align: start;
+  }
+
+  .form {
+    width: 100%;
+  }
+
+  .form input {
+    width: 60%;
+    font-size: 1.2rem;
+  }
+
+  .form select {
+    width: 20%;
+  }
+
+  .submit-btn {
+    width: 15%;
+  }
+
+  .categories h2 {
+    font-size: 2rem;
+  }
+}
+
+@media all and (min-width: 1025px) {
+  .progress-chapter {
+    gap: 80px;
+  }
+}
